Add top level and sidebar navigation to development-guide
In order to allow navigation to files (pages) inside the
development-pratices folder an index.md file was created that links to the
top bar navigation. This page acts only as a landing page in which the
side bar navigation can be displayed for the development guides.

The user workflow is as follows:
1. The user lands at the home page
2. The user clicks Development Pratices in the top bar navigation
3. The user is presented with a basic page that displays all of the
development pratices in the side bar navigation